Break apart worker into more modules
