+++ analysis/rekognition_worker.py
@@ -1,16 +1,14 @@
-import boto3
 import botocore.exceptions
 import logging as log
 import time
-import threading
-import json
 import concurrent.futures
 import enum
 import worker.settings as settings
 from functools import partial
-from json.decoder import JSONDecodeError
 from collections import defaultdict, Counter
 from confluent_kafka import Consumer, Producer
+from analysis.task import handle_image_task
+from analysis.util import LocalTokenBucket, RecentlyProcessed, parse_msg
 
 IMG_BUCKET = 'cc-image-analysis'
 LABELS_TOPIC = 'image_analysis_labels'
@@ -29,102 +27,6 @@ class TaskStatus(enum.Enum):
     SUCCEEDED = 1
     IGNORED_DUPLICATE = 2
     ERROR = 3
-
-
-class LocalTokenBucket:
-    """
-    A thread-safe token bucket for ensuring conformance to a rate limit.
-    """
-    def __init__(self, max_val, refresh_rate_sec=1):
-        self._refresh_rate_sec = refresh_rate_sec
-        self._lock = threading.Lock()
-        self._last_timestamp = None
-        self._MAX_TOKENS = max_val
-        self._curr_tokens = max_val
-
-    def _acquire_token(self):
-        with self._lock:
-            now = time.time()
-            if not self._last_timestamp:
-                self._last_timestamp = now
-            elif now - self._last_timestamp > self._refresh_rate_sec:
-                # Replenish
-                self._curr_tokens = self._MAX_TOKENS
-                self._last_timestamp = now
-            if self._curr_tokens <= 0:
-                return False
-            self._curr_tokens -= 1
-            return True
-
-    def throttle_fn(self, task_to_throttle):
-        """
-        Wrap the input function and make it block until it acquires a token from
-        the bucket instance. (Thread-safe)
-        """
-        token_acquired = False
-        while not token_acquired:
-            token_acquired = self._acquire_token()
-            time.sleep(0.01)
-        return task_to_throttle()
-
-
-class RecentlyProcessed:
-    """ Determine whether an item has been seen recently. (Thread-safe) """
-    def __init__(self, retention_num):
-        # For fast membership checks
-        self._set = set()
-        # For remembering order of arrival
-        self._queue = []
-        self._max_retain = retention_num
-        self._lock = threading.Lock()
-
-    def seen_recently(self, item):
-        with self._lock:
-            if item in self._set:
-                return True
-            self._set.add(item)
-            self._queue.append(item)
-            if len(self._queue) > self._max_retain:
-                to_delete = self._queue.pop(0)
-                self._set.remove(to_delete)
-            return False
-
-
-def detect_labels_query(image_uuid, boto3_session):
-    img = {
-        'S3Object': {
-            'Bucket': 'cc-image-analysis',
-            'Name': f'{image_uuid}.jpg'
-        }
-    }
-    rekog_client = boto3_session.client('rekognition')
-    response = rekog_client.detect_labels(Image=img)
-    data = {
-        'image_uuid': image_uuid,
-        'response': response
-    }
-    return data
-
-
-def enqueue(rekognition_response: dict, kafka_producer):
-    resp_json = json.dumps(rekognition_response).encode('utf-8')
-    kafka_producer.produce(LABELS_TOPIC, resp_json)
-
-
-def handle_image_task(message, output_producer, recent_ids):
-    """
-    Get Rekognition labels for an image and output results to a Kafka topic.
-
-    Only call Rekognition if we haven't recently processed the ID.
-    """
-    msg = json.loads(message)
-    image_uuid = msg['identifier']
-    if recent_ids.seen_recently(image_uuid):
-        return TaskStatus.IGNORED_DUPLICATE
-    boto3_session = boto3.session.Session()
-    response = detect_labels_query(image_uuid, boto3_session)
-    enqueue(response, output_producer)
-    return TaskStatus.SUCCEEDED
 
 
 def _monitor_futures(futures):
@@ -148,15 +50,6 @@ def _monitor_futures(futures):
     return _futures, Counter(statuses)
 
 
-def _parse_msg(msg):
-    try:
-        parsed = json.loads(str(msg.value(), 'utf-8'))
-        return parsed['identifier']
-    except (KeyError, JSONDecodeError):
-        log.warning(f'Failed to parse message {msg}')
-        return None
-
-
 def _poll_work(msg_buffer, msgs_remaining, consumer):
     """ Poll consumer for messages and parse them. """
     while len(msg_buffer) < NUM_MESSAGES_BUFFER and msgs_remaining:
@@ -165,7 +58,7 @@ def _poll_work(msg_buffer, msgs_remaining, consumer):
             log.info('No more messages remaining')
             msgs_remaining = False
         else:
-            parsed = _parse_msg(msg)
+            parsed = parse_msg(msg)
             if parsed:
                 msg_buffer.append(parsed)
     return msgs_remaining
@@ -210,16 +103,17 @@ def listen(consumer, producer, task_fn):
     executor.shutdown(wait=True)
     _, future_stats = _monitor_futures(futures)
     status_tracker += future_stats
-    log.info(f'Total stats: {status_tracker}')
+    log.info(f'Aggregate stats: {status_tracker}')
     log.info('Worker shutting down')
 
 
 if __name__ == '__main__':
     log.basicConfig(level=log.INFO, format='%(asctime)s %(message)s')
     output_producer = Producer({'bootstrap.servers': settings.KAFKA_HOSTS})
-    consumer = Consumer({
+    inbound_consumer = Consumer({
         'bootstrap.servers': settings.KAFKA_HOSTS,
-        'group.id': 'rekognition_worker',
+        'group.id': 'image_metadata_updates',
         'auto.offset.reset': 'earliest'
     })
-    listen(consumer, output_producer, handle_image_task)
+    inbound_consumer.subscribe([])
+    listen(inbound_consumer, output_producer, handle_image_task)
